Redid keymaps to conserve function keys, made kill-buffer-shortly function, on \C-f4.
Behavior of kill-buffer-shortly is to turn on tempbuf-mode, and then bury the buffer.
That way, it is still around if we really want it. But otherwise, it will be killed.

Current issue: want to turn tempbuff-mode off next time we switch to it.

darcs-hash:20070504230802-00070-430a2fb671228b4f662b5b546d44d8b8ae768b72.gz