Rebuilt darcsum-ediff.
Now, upon closing the ediff-buffer, it switches back to the old window configuration.
Also, the darcs pristine version of the file (located in _darcs/pristine) is opened
as READ ONLY. Additionally, turns on tempbuf-mode in the pristine buffer, so that it
dies when not needed.

darcs-hash:20070517201954-00070-0e3e34719735cbbb9ecf96b3694d9893a30647dc.gz