Added command 'find-file-temporary.
Opens a file, but in the style of an info-document. I.e., it opens it up in
another buffer, doesn't switch to the window, and simply splits the screen if
need be. The new buffer is a temporary buffer, and will die soon if it is not
made permanent.

darcs-hash:20070604181136-00070-985665d8f9f20466912ffec17a8ef586efe2fc3d.gz